Added JWT Authentication & Authorisation

+++ workboard-backend/src/main/java/com/project/workboard/repository/BoardListRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.project.workboard.entity.BoardList;
 
 @Repository
-public interface BoardListRepository extends JpaRepository<BoardList, Long>{
+public interface BoardListRepository extends JpaRepository<BoardList, Integer>{
 
 }
