+++ workboard-backend/src/main/java/com/project/workboard/repository/TaskCardRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.project.workboard.entity.TaskCard;
 
 @Repository
-public interface TaskCardRepository extends JpaRepository<TaskCard, Long>{
+public interface TaskCardRepository extends JpaRepository<TaskCard, Integer>{
 
 }
